@uppy/drag-drop - moved isDragDropSupported() to @uppy/utils
