+++ packages/@uppy/utils/src/isDragDropSupported.js
@@ -1,0 +1,22 @@
+/**
+ * Checks if the browser supports Drag & Drop (not supported on mobile devices, for example).
+ *
+ * @return {Boolean}
+ */
+module.exports = function isDragDropSupported () {
+  const div = document.createElement('div')
+
+  if (!('draggable' in div) || !('ondragstart' in div && 'ondrop' in div)) {
+    return false
+  }
+
+  if (!('FormData' in window)) {
+    return false
+  }
+
+  if (!('FileReader' in window)) {
+    return false
+  }
+
+  return true
+}
