+++ packages/@uppy/utils/types/index.d.ts
@@ -93,6 +93,10 @@ declare module '@uppy/utils/lib/isObjectURL' {
   export default function isObjectURL(url: string): boolean;
 }
 
+declare module '@uppy/utils/lib/isDragDropSupported' {
+  export default function isDragDropSupported(): boolean;
+}
+
 declare module '@uppy/utils/lib/isPreviewSupported' {
   export default function isPreviewSupported(mime: string): boolean;
 }
